Replace `fetchPatientPhotoUrl` with `usePatientPhoto` hook
Replace the fetchPatientPhotoUrl implementation with the usePatientPhoto SWR hook. This has the added advantage of fixing a pretty gnarly bug in the patient registration component. Currently, when the registration form is launched in edit mode, `fetchPatientPhotoUrl` gets invoked conditionally in a useEffect hook. This useEffect hook has the patient and config objects as part of its dependency array. If you open the network tab in the devtools console with the form open, you'll notice that the `fetchPatientPhotoUrl` runs indefinitely. Leave it open long enough and your browser tab will certainly crash from a stack overflow. I'm not sure if this fix merits it's own PR or if it's worth leaving it in as part of O3-940.