O3-940: Redesign patient avatars (#68)
* O3-940: Redesign patient avatars

Redesign patient avatars to match the designs outlined in https://app.zeplin.io/project/61434fa756474d5545f65cf4/screen/614362b21dd66d20a276d8b8.

Summary of the work done:

- Install `react-avatar`.
- Install `geopattern`.
- Add a `usePatientPhoto` hook that leverages SWR to fetch the patient photo (nice, but not strictly required).
- Refactor DisplayPhotoComponent to use the `Avatar` component from react-avatar. Pass the patient name in as the value of the name prop. This allows the Avatar component to derive the patient's initials and display that as a fallback when an image source is not provided. Set the colour prop to a transparent colour so that the background image (provided by geopattern can be seen). Set the `photo.imageSrc` value from the aforementioned `usePatientPhoto` hook as the src prop. Finally, set the background image of the avatar to the pattern generated by geopattern using the patient UUID.

* Replace `fetchPatientPhotoUrl` with `usePatientPhoto` hook

Replace the fetchPatientPhotoUrl implementation with the usePatientPhoto SWR hook. This has the added advantage of fixing a pretty gnarly bug in the patient registration component. Currently, when the registration form is launched in edit mode, `fetchPatientPhotoUrl` gets invoked conditionally in a useEffect hook. This useEffect hook has the patient and config objects as part of its dependency array. If you open the network tab in the devtools console with the form open, you'll notice that the `fetchPatientPhotoUrl` runs indefinitely. Leave it open long enough and your browser tab will certainly crash from a stack overflow. I'm not sure if this fix merits it's own PR or if it's worth leaving it in as part of O3-940.