(fix) Move data-floating-menu-container attribute to DataTable tag
This PR moves the `data-floating-menu-container` attribute from the container div to the DataTable tag. Carbon uses this attribute to determine where to render the OverflowMenu component. Presently, the OverflowMenu gets rendered outside the bounds of the DataTable content area. This change fixes that issue.

Unrelatedly, this PR also removes an unused Carbon component import.