(chore) Add webpack as an explicit dev dependency
Presently, developers are not able to run webpack-dependent scripts such as `yarn serve` in this repository. Following our recent migration to Yarn v3, we are now required to explicitly state dev dependencies. To that end,this PR adds webpack as an dev dependency to all the packages in this repository.