(fix) Fix tests for the Date of Birth component
This PR provides a bunch of anticipatory fixes necessary to support a tweak to the registration form. More specifically, @makombe wishes to change the type of `datePickerType` prop from `simple` to `single`. The big benefit from that would be the provision of a calendar UI from which users can more conveniently select the date of the birth value. Applying that change with the current setup yields a test failure. This failure is caused by the react module mock in `dob.test.tsx`. More specifically, that mock went on to provide a partial `useContext` mock. This wound up breaking some of the Carbon date picker component functionality that depends on `useContext`.

This PR gets rid of module mocks for React and Formik in the `DobComponent` test suite and refactors the `render` function to return a component wrapped in a `PatientRegistrationContext`. This refactor enables the test to pass both with the current setup as well as with @makombe's anticipated change.