Remove initialSelectedItems from service type multi-select
Makes it so that the multi-select is not pre-selected with the first service type. With this change, no options are pre-selected by default. This more accurately wires up the filter to default to a blank state where all available appointments are shown. Users will have to explicitly choose which service types they want to filter by, rather than having one pre-selected for them.