(refactor) Remove custom modal header close button styles (#1541)
This commit removes the custom modal header close button styles from modals in this
repository. This is necessitated by removal of the custom logic in the modal system that
manually injected the close button into the modal header following the recent migration to
v1.76.0 of the Carbon Design System in Core.